requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
firebase-admin==6.8.0
stripe==12.1.0
python-dotenv==1.0.1